pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ttt_theory"
version = "1.0.0"
description = "Experiments for the theory of TTT paper"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["machine learning", "deep learning", "transformers", "theory"]
dependencies = [
    # Core scientific computing
    "numpy~=2.2.6",
    "scipy~=1.15.3",
    "pandas~=2.3.0",

    # Machine Learning
    "torch~=2.7.1",
    "torchvision==0.22.1",
    "scikit-learn==1.6.1",
    "faiss-cpu==1.10.0",

    # Transformers and NLP
    "transformers~=4.52.4",
    "datasets~=2.21.0",
    "openai-clip~=1.0",

    # Visualization
    "matplotlib~=3.10.3",
    "seaborn~=0.13.2",

    # Utilities
    "tqdm~=4.67.1",
    "tabulate~=0.9.0",
    "pyyaml~=6.0.2",
    "requests~=2.32.3",
    "pillow~=11.2.1",
    "pydantic~=2.11.5",

    # Configuration and experiment management
    "hydra-core~=1.3.2",
    "hydra-optuna-sweeper~=1.2.0",
    "hydra-joblib-launcher~=1.2.0",
    "omegaconf~=2.3.0",
    "wandb~=0.20.1",

    # Jupyter
    "ipywidgets==8.1.5",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["sae", "scaling"]